requirements.txt
# Core dependencies
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
termcolor>=2.0.0
python-dotenv>=1.0.0
commentjson>=1.2.0

# Voice processing
piper-tts>=1.2.0
faster-whisper>=1.0.0
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0

# MCP client
mcp>=1.0.0
aiosqlite>=0.19.0

# Optional but recommended
psutil>=5.9.0
